+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unreleased
+### Added
+* Added support for specular color maps to generated materials.
+
+### Fixed
+* Fixed an issue where glossiness would use the incorrect material assignment.
+
 ## 0.9.1 - 2024-06-24
 ### Fixed
 * Fixed an issue where importing would fail if the UI language was not set to English.